Готовы окна для действий с сотрудником
- Редактирование и добавление полностью работают для взаимодействия с данными сотрудников
- Разделены стили комбобоксов для вывода данных (со свойством Name и без него)
- Добавлены сервисы для сотрудника
- Добавлены сообщения для ограничения ввода данных (для сотрудника)

+++ WPF-CourierFrim/WPF-CourierFrim/Classes/Validators.cs
@@ -179,7 +179,7 @@ namespace WPF_CourierFrim.Classes
         // Ограничение по возрасту 
         public static bool ValidateCorrectAge(DateOnly date)
         {
-            short age = TypeHelper.CalculateAge(date);
+            int age = TypeHelper.CalculateAge(date);
 
             if (age < 18 || age > 80)
             {
